:racehorse: Reduce allocation in NRLMSISE-00
The Legendre matrix used in the original NRLMSISE-00 is upper
triangular, whereas we use the lower triangular representation. Thus, we
needed to transpose the matrix to keep the algorithm implemented here
close to the original. However, the code was allocating a new matrix.
Now, we just store the `Adjoint` pointer, which reduced one allocation
and improve the performance.

We also lower the requirement for the pre-allocated Legendre matrix size
since we only use terms with degree 8 and order 4 at maximum.